Update Banyak Fitur dan Perbaikan

+++ database/migrations/2026_07_11_084232_add_new_columns_to_pengajuan_magangs_table.php
@@ -11,20 +11,20 @@ return new class extends Migration
      */
 public function up(): void
 {
-    Schema::table('pengajuan_magangs', function (Blueprint $table) {
-
-        $table->text('catatan')
-            ->nullable()
-            ->after('status');
-
-    });
+    if (! Schema::hasColumn('pengajuan_magangs', 'catatan')) {
+        Schema::table('pengajuan_magangs', function (Blueprint $table) {
+            $table->text('catatan')
+                ->nullable()
+                ->after('status');
+        });
+    }
 }
 
 public function down(): void
 {
-    Schema::table('pengajuan_magangs', function (Blueprint $table) {
-
-        $table->dropColumn('catatan');
-
-    });
+    if (Schema::hasColumn('pengajuan_magangs', 'catatan')) {
+        Schema::table('pengajuan_magangs', function (Blueprint $table) {
+            $table->dropColumn('catatan');
+        });
+    }
 }};
